Large upload bug fixes

+++ app/api/db/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, LargeBinary, DateTime
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, BigInteger, String, LargeBinary, DateTime
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from db.session import Base
@@ -59,7 +59,7 @@ class Mod(Base):
     modDescription = Column(String(1200))
     modAuthor = Column(String(128), index=True)
     modFileName = Column(String(128), index=True)
-    modFileSize = Column(Integer, index=True)
+    modFileSize = Column(BigInteger, index=True)
     modId = Column(Integer, index=True)
     mod_downloads = Column(Integer, index=True)
     modTags = Column(String(64))
